Allow for optionally passing in a pre-existing VPC + Remove Clickhouse refs. (#121)
Users can now optionally disable creation of the main VPC that hosts the
braintrust services and data stores. An existing VPC and subnets can be
passed in instead. By default, the module will continue to create a VPC
as it has always.

To disable creation of the main VPC:

```hcl
module "braintrust-data-plane" {
  source = "github.com/braintrustdata/terraform-braintrust-data-plane"
  # ... your existing configuration ...
  # Use existing VPC
  create_vpc = false
  existing_vpc_id                        = "vpc-xxxxxxxxx"
  existing_private_subnet_1_id           = "subnet-xxxxxxxxx"
  existing_private_subnet_2_id           = "subnet-yyyyyyyyy"
  existing_private_subnet_3_id           = "subnet-zzzzzzzzz"
  existing_public_subnet_1_id            = "subnet-aaaaaaaaa"
}
```

There were old references to Clickhouse that were removed. No existing
TF users should be on Clickhouse. We dropped support for it earlier in
the year. This change removes variables, outputs, and the references to
the module that was already disabled.