Allow user to optionally pass in database subnet ids (#124)
Normally the RDS instance uses the private subnets from the VPC, but
some customers want to be able to specify a custom VPC with custom
subnets for the DB. The new variable `database_subnet_ids` allows them
to do that. By default it will just use the private subnets from the
main VPC exactly as it always has.

```
module "braintrust-data-plane" {
  source = "github.com/braintrustdata/terraform-braintrust-data-plane"
  # ... your existing configuration ...
  # Use existing VPC
  create_vpc = false
  existing_vpc_id                        = "vpc-xxxxxxxxx"
  existing_private_subnet_1_id           = "subnet-xxxxxxxxx"
  existing_private_subnet_2_id           = "subnet-yyyyyyyyy"
  existing_private_subnet_3_id           = "subnet-zzzzzzzzz"
  existing_public_subnet_1_id            = "subnet-aaaaaaaaa"

  database_subnet_ids = ["subnet-ddddddddd", "subnet-eeeeeeeee"]
}
```