Add state rollback support for v3 back to v2 (#140)
The new v3.0.0 of terraform involves a lot of state movement. In the
event a user has to revert back from v3 to v2, they will need their
state to be placed back into the original locations.

This code is basically the opposite of `moved_state.tf` in the v3.0.0
release.

This file will have no effect on existing users. If it can't find state
in the `from =` block it just ignores it.

## Testing
* Created a fresh stack on latest v2 with this file included
* Pointed module to the v3.0.0 branch and applied
* Resources were moved to new locations and minor expected modifications
were made
* Pointed module back to this branch (Simulating a rollback to v2)
* Resources were moved back to original locations and minor expected
changes from v3 were reverted
* Commented out the rollback_v2_state.tf file
* Plan still shows no changes as expected
* Pointed module back to v3.0.0 branch and applied
* Resources were moved to new locations and minor expected modifications
were made