Revert to support OpenTofu 1.10.0 (#156)
A customer is using OpenTofu and unfortunately tofu lags significantly
behind official terraform. This PR reverts the requirement to use
terraform >= 1.12.0. OpenTofu is only on release 1.10 at the moment
which is very unfortunate.

We are now essentially required to use the least common denominator of
features and bug fixes that exist with Terraform and OpenTofu. This
blocks us from being able to use ephemeral values to keep secrets out of
state, which is a common customer request (and a potential blocker in
the future). Tofu 1.11 is likely to release soon with initial support
for it so there is some hope there.

Changes:
* Tofu does not have fixes for null checks validation blocks so I had to
remove them. These were added from direct customer complaints, but we
have no choice.
* Added tofu to the basic CI validation checks. 
* Note: we need to start doing actual plans in CI since some
incompatibilities are only known at plan time.
* Added tofu to mise so we can test planning locally