Upgrade AWS provider to v6 (#161)
> [!WARNING]
> It is not possible to roll back after upgrading to this version. The
new AWS provider modifies your state and it will not be compatible with
v5. It is recommended to upgrade to our latest
[v3.1.3](https://github.com/braintrustdata/terraform-aws-braintrust-data-plane/releases/tag/v3.1.3)
terraform module before upgrading to v4.0.0 which includes this change.

The AWS v6 provider has support for passing in region as an attribute
rather than having to use multiple providers. Additionally it has
support for setting Cloudfront to use TLS1.3 as the required protocol
with modern cipher suites.