Explicitly fill out the source_arn for lambda permissions (#162)
There appears to be some strange bug in the AWS provider that leaves the
AccountId part of the API Gateway execution arn empty. It doesn't happen
to 99% of our users but consistently hits at least two of them now.

Instead of relying on the `.execution_arn` attribute of the API gateway
the ARN is now manually constructed using the current account ID and
region configured in the provider. This should be backwards compatible
for existing users and our test stacks show zero diff in the terraform
plan.