database: use kms_key_arn_db only for the database
Without this patch we get the following failure because the cmk we use
with kms for rds may not also be used for ssm.

```
│ Error: creating Secrets Manager Secret (a209178-prod-brain/DatabaseSecret-20251204183547822800000008): operation error Secrets Manager: CreateSecret, https response error StatusCode: 400, RequestID: 8081dad5-b5c5-4a3e-b701-f4e67aba1125, api error AccessDeniedException: Access to KMS is not allowed
│
│   with module.braintrust-data-plane.module.database.aws_secretsmanager_secret.database_secret,
│   on .terraform/modules/braintrust-data-plane/modules/database/main.tf line 153, in resource "aws_secretsmanager_secret" "database_secret":
│  153: resource "aws_secretsmanager_secret" "database_secret" {
```

This patch resolves the error by using kms_key_arn_db only for the
aws_db_instance resource.  The kms_key_arn is used for the
aws_secretmanager_secret.database_secret resource. The kms_key_arn
corresponds to the vendor managed KMS resource managed by the Braintrust
terraform module