Add BRAINSTORE_AI_PROXY_URL to Brainstore (#157)
Brainstore needs to be able to talk to the AI Proxy and so it needs to
discover the Proxy URL. Unfortunately this creates a cycle when first
initializing a data plane from scratch. The only way to address this is
to allow Brainstore in TF to be created but fail to start up until it is
able to discover the AI Proxy URL from AWS SSM parameter store. The AI
Proxy will publish its URL once it has completed.

Scenario: New stack from scratch
1. Brainstore is created and goes into a launch failure loop since it
can't find the AI Proxy URL
2. AI proxy and function URL is created
3. AI Proxy Function URL is stored in SSM
4. Brainstore ASG creates a new node that discovers the URL and
successfully starts

Scenario: Existing stack updating
1. New Brainstore instances are created but potentially go into a launch
failure loop since it can't find the AI Proxy URL
2. AI Proxy Function URL is stored in SSM (this likely happens so fast
that the first item doesn't happen)
3. Brainstore ASG creates a new node and discovers the URL and
successfully starts
4. Old brainstore nodes are cycled out

---------

Co-authored-by: Mike Deeks <mike@braintrustdata.com>